Core Foundation
Restructuring of Phil's branch to provide a more flexible code base. It
has many additions. The additions include:
-Implementation of FSM
-Implementation of two states; MainMenu and Game which are operational
-Implementation of a text rendering class Message
-Implementation of an image rendering class
-Implementation of an animation class

The functionality implemented is in it's infancy and needs tweaking
improvements and testing.

A number of bugs have also been pulled from Phil's branch including an
array out of bounds error when the mouse leaves the screen sometimes and
when the game goes full screen. The game can't cope with a resolution
bigger than what's currently defined.